+++ examples/mqtt/mqtt_example.rs
@@ -68,7 +68,7 @@ mqtt_incoming_data_cb(arg: &mut Vec<u8>, const u8 *data, len: u16, flags: u8)
   LWIP_UNUSED_ARG(data);
 
   printf("MQTT client \"%s\" data cb: len %d, flags %d\n", client_info.client_id,
-          (int)len, (int)flags);
+          len, flags);
 }
 
 pub fn
@@ -77,7 +77,7 @@ mqtt_incoming_publish_cb(arg: &mut Vec<u8>, const char *topic, tot_len: u32)
   const struct mqtt_connect_client_info_t* client_info = (const struct mqtt_connect_client_info_t*)arg;
 
   printf("MQTT client \"%s\" publish cb: topic %s, len %d\n", client_info.client_id,
-          topic, (int)tot_len);
+          topic, tot_len);
 }
 
 pub fn
@@ -85,7 +85,7 @@ mqtt_request_cb(arg: &mut Vec<u8>, err: err_t)
 {
   const struct mqtt_connect_client_info_t* client_info = (const struct mqtt_connect_client_info_t*)arg;
 
-  printf("MQTT client \"%s\" request cb: err %d\n", client_info.client_id, (int)err);
+  printf("MQTT client \"%s\" request cb: err %d\n", client_info.client_id, err);
 }
 
 pub fn
@@ -94,7 +94,7 @@ mqtt_connection_cb(mqtt_client_t *client, arg: &mut Vec<u8>, mqtt_connection_sta
   const struct mqtt_connect_client_info_t* client_info = (const struct mqtt_connect_client_info_t*)arg;
   LWIP_UNUSED_ARG(client);
 
-  printf("MQTT client \"%s\" connection cb: status %d\n", client_info.client_id, (int)status);
+  printf("MQTT client \"%s\" connection cb: status %d\n", client_info.client_id, status);
 
   if (status == MQTT_CONNECT_ACCEPTED) {
     mqtt_sub_unsub(client,
