+++ src/api/netdb.rs
@@ -373,7 +373,7 @@ pub fn lwip_getaddrinfo(const char *nodename, const char *servname,
     inet6_addr_from_ip6addr(&sa6.sin6_addr, ip_2_ip6(&addr));
     sa6.sin6_family = AF_INET6;
     sa6.sin6_len = sizeof(struct sockaddr_in6);
-    sa6.sin6_port = lwip_htons((u16)port_nr);
+    sa6.sin6_port = lwip_htons(port_nr);
     sa6.sin6_scope_id = ip6_addr_zone(ip_2_ip6(&addr));
     ai.ai_family = AF_INET6;
 
@@ -384,7 +384,7 @@ pub fn lwip_getaddrinfo(const char *nodename, const char *servname,
     inet_addr_from_ip4addr(&sa4.sin_addr, ip_2_ip4(&addr));
     sa4.sin_family = AF_INET;
     sa4.sin_len = sizeof(struct sockaddr_in);
-    sa4.sin_port = lwip_htons((u16)port_nr);
+    sa4.sin_port = lwip_htons(port_nr);
     ai.ai_family = AF_INET;
 
   }
